feat(payments): centralize status info management by moving getStatusInfo to utils
